Feature/accept reject estimate

+++ src/app/modules/estimates/List/container.jsx
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import List from './components/list';
 import Estimate from '../model';
 import EstimateController from '../controller';
+import UserController from '../../user/controller';
 
 class EstimateListContainer extends React.Component {
   constructor(props, context) {
@@ -10,22 +11,53 @@ class EstimateListContainer extends React.Component {
     this.store = context.store;
     this.state = {
       estimates: this.props.estimates,
+      user: UserController.currentUser(this.store),
     };
 
     this.onClickEdit = this.onClickEdit.bind(this);
     this.onClickDelete = this.onClickDelete.bind(this);
+    this.onClickAccept = this.onClickAccept.bind(this);
+    this.onClickReject = this.onClickReject.bind(this);
   }
 
   componentWillReceiveProps(nextProps) {
     this.setState({ estimates: nextProps.estimates });
   }
 
+  onClickAccept(estimate) {
+    if (estimate.state !== 'accepted') {
+      return () => {
+        EstimateController.acceptEstimate(this.store, estimate);
+      };
+    }
+    return undefined;
+  }
+
+  onClickReject(estimate) {
+    if (estimate.state !== 'rejected') {
+      return () => {
+        EstimateController.rejectEstimate(this.store, estimate);
+      };
+    }
+    return undefined;
+  }
+
   onClickEdit(estimate) {
-    EstimateController.editEstimate(this.store, estimate);
+    if (this.state.user.id === estimate.user_id && estimate.state !== 'accepted') {
+      return () => {
+        EstimateController.editEstimate(this.store, estimate);
+      };
+    }
+    return undefined;
   }
 
   onClickDelete(estimate) {
-    EstimateController.deleteEstimate(this.store, estimate);
+    if (this.state.user.id === estimate.user_id) {
+      return () => {
+        EstimateController.deleteEstimate(this.store, estimate);
+      };
+    }
+    return undefined;
   }
 
   render() {
@@ -34,6 +66,8 @@ class EstimateListContainer extends React.Component {
         estimates={this.state.estimates}
         onClickEdit={this.onClickEdit}
         onClickDelete={this.onClickDelete}
+        onClickAccept={this.onClickAccept}
+        onClickReject={this.onClickReject}
       />
     );
   }
